Added new main activity and minor changes
Start up activity is now a page containing two buttons that link to
login and sign up respectively. Sign up now converts all usernames to
lowercase before sending them to the database. Also checks for spaces.
Login now converts all usernames to lowercase before checking with the
server. This should make it so that swagmaster and SwAgMaSTeR are the
same username.